move helpers functions from fetch to utils
