+++ src/fetch/mod.rs
@@ -2,7 +2,6 @@ use serde::{Deserialize, Serialize};
 use wasm_bindgen::prelude::wasm_bindgen;
 
 pub mod fetch_api;
-pub mod formdata;
 mod req_properties;
 
 #[wasm_bindgen(getter_with_clone)]
